Reworks middleware to use utils
Adds capturing of request and response into utils functions. Lets the
middleware use the utils as well.
